// Fortune.Web/src/api/FortuneApiClient.ts
export class FortuneModel implements IFortuneModel {
    id?: string;
    longFortune?: string | undefined;
    shortFortune?: string | undefined;
    imageTopics?: string | undefined;
    qrImage?: string | undefined;
    audio?: string | undefined;
    fortuneImage?: string | undefined;
    luckyNumbers?: number[] | undefined;
    fortuneUsed?: boolean;
    fortuneType?: string | undefined;

    constructor(data?: IFortuneModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.longFortune = _data["longFortune"];
            this.shortFortune = _data["shortFortune"];
            this.imageTopics = _data["imageTopics"];
            this.qrImage = _data["qrImage"];
            this.audio = _data["audio"];
            this.fortuneImage = _data["fortuneImage"];
            if (Array.isArray(_data["luckyNumbers"])) {
                this.luckyNumbers = [] as any;
                for (let item of _data["luckyNumbers"])
                    this.luckyNumbers!.push(item);
            }
            this.fortuneUsed = _data["fortuneUsed"];
            this.fortuneType = _data["fortuneType"];
        }
    }

    static fromJS(data: any): FortuneModel {
        data = typeof data === 'object' ? data : {};
        let result = new FortuneModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["longFortune"] = this.longFortune;
        data["shortFortune"] = this.shortFortune;
        data["imageTopics"] = this.imageTopics;
        data["qrImage"] = this.qrImage;
        data["audio"] = this.audio;
        data["fortuneImage"] = this.fortuneImage;
        if (Array.isArray(this.luckyNumbers)) {
            data["luckyNumbers"] = [];
            for (let item of this.luckyNumbers)
                data["luckyNumbers"].push(item);
        }
        data["fortuneUsed"] = this.fortuneUsed;
        data["fortuneType"] = this.fortuneType;
        return data;
    }
}

export interface IFortuneModel {
    id?: string;
    longFortune?: string | undefined;
    shortFortune?: string | undefined;
    imageTopics?: string | undefined;
    qrImage?: string | undefined;
    audio?: string | undefined;
    fortuneImage?: string | undefined;
    luckyNumbers?: number[] | undefined;
    fortuneUsed?: boolean;
    fortuneType?: string | undefined;
}